+++ extension/tuning/package.nls.en.json
@@ -1,6 +1,7 @@
 {
     "perfadvisor.configer.server": "You are not connected to the remote server.\n[Configure Server](command:extension.view.perfadvisorserverconfig)",
     "perfadvisor.again.configer.server": "You are connected to the remote server\n[Reconfigure the server](command:extension.view.perfadvisorserverconfigagain)",
+    "perfadvisor.server.refresh.failure": "Failed to connect to the remote server.\n[Configure Server](command:extension.view.perfadvisorserverconfig)",
 
     "perfadvisor.free.trial": "You can also apply for a Kunpeng remote lab for trial. A remote lab is pre-installed with the Kunpeng Porting Advisor, Kunpeng Compiler, Kunpeng Hyper Tuner, and Dynamic Binary Translator (ExaGear).\n[Free Trial](command:extension.view.perfadvfreetrialremoteenvironment)",
     "perfadvisor.no.login": "You are not logged in.\n[Log In](command:extension.view.perfadvisorlogin)",
