+++ extension/tuning/package.nls.json
@@ -1,6 +1,7 @@
 {
-    "perfadvisor.configer.server": "您还未连接远端服务器。\n[配置远程服务器](command:extension.view.perfadvisorserverconfig)",
+    "perfadvisor.configer.server": "您还未连接远端服务器。\n[配置远端服务器](command:extension.view.perfadvisorserverconfig)",
     "perfadvisor.again.configer.server": "您已连接远端服务器。\n[重新配置服务器](command:extension.view.perfadvisorserverconfigagain)",
+    "perfadvisor.server.refresh.failure": "远端服务器连接失败。\n[配置远端服务器](command:extension.view.perfadvisorserverconfig)",
 
     "perfadvisor.free.trial": "您也可申请免费试用鲲鹏远程实验室，该环境已预装鲲鹏代码迁移工具、鲲鹏编译器、鲲鹏性能分析工具、动态二进制翻译工具（ExaGear）。\n[免费试用](command:extension.view.perfadvfreetrialremoteenvironment)",
     "perfadvisor.no.login": "您还未登录。\n[登录](command:extension.view.perfadvisorlogin)",
