+++ extension/tuning/src/extension/media/sidebar.css
@@ -2,41 +2,7 @@ body {
     background-color: transparent;
 }
 
-.color-list {
-    list-style: none;
-    padding: 0;
+.info-box{
+    border-collapse:separate;
+    border-spacing:0px 16px;
 }
-
-.color-entry {
-    width: 100%;
-    display: flex;
-    margin-bottom: 0.4em;
-    border: 1px solid var(--vscode-input-border);
-}
-
-.color-preview {
-    width: 2em;
-    height: 2em;
-}
-
-.color-preview:hover {
-    outline: inset white;
-}
-
-.color-input {
-    display: block;
-    flex: 1;
-    width: 100%;
-    color: var(--vscode-input-foreground);
-    background-color: var(--vscode-input-background);
-    border: none;
-    padding: 0 0.6em;
-}
-
-.add-color-button {
-    display: block;
-    border: none;
-    margin: 0 auto;
-}
-
-
